some vulture hints
